Graph validator debugs and verifies all computational graph related issues

+++ arena.h
@@ -1,6 +1,8 @@
 #ifndef ARENA_H
 #define ARENA_H
 
+#include <stddef.h>
+
 typedef struct Arena {
 	char *base;
 	size_t size;
@@ -8,9 +10,9 @@ typedef struct Arena {
 } Arena;
 
 // prototype definitions
-void arena_init(Arena *A, int size);
+void arena_init(Arena *A, size_t size);
 void arena_reset(Arena *A);
-void *arena_alloc(Arena *A, int size);
+void *arena_alloc(Arena *A, size_t size);
 
 
 #endif
